new:endpoint
CRUD Listo para <storages>
Se debe continuar con items

+++ app/blueprints/api_v1/storages.py
@@ -10,12 +10,12 @@ from app.utils.exceptions import APIException
 from app.utils.helpers import JSONResponse, pagination_form
 from app.utils.decorators import json_required, user_required
 from app.utils.db_operations import get_user_by_id
-from app.utils.validations import only_letters, validate_inputs
+from app.utils.validations import validate_inputs, validate_string
 
 
 storages_bp = Blueprint('storages_bp', __name__)
 
-@storages_bp.route('/', methods=['GET', 'DELETE'])
+@storages_bp.route('/', methods=['GET', 'PUT', 'DELETE'])
 @json_required()
 @user_required()
 def get_storages():
@@ -30,10 +30,8 @@ def get_storages():
         raise APIException('invalid format in query string, <int> is expected')
 
     if storage_id == -1:
-        
         if request.method == 'GET':
-
-            s = user.company.storages.order_by(Storage.id.asc()).paginate(page, per_page) #return all storages,
+            s = user.company.storages.order_by(Storage.name.asc()).paginate(page, per_page) #return all storages,
             return JSONResponse(
                 message="ok",
                 payload={
@@ -41,8 +39,9 @@ def get_storages():
                     **pagination_form(s)
                 }
             ).to_json()
-
-        abort(405)
+    
+        if request.method == 'DELETE': #! can't delete in bulk...
+            raise APIException("missing <storage-id> in query string to delete")
 
     #if an id has been passed in as a request arg.
     s = user.company.storages.filter(Storage.id == storage_id).first()
@@ -50,8 +49,7 @@ def get_storages():
         raise APIException(f"storage-id-{storage_id} not found", status_code=404, app_result="not_found")
 
     if request.method == 'GET': 
-    #return the storage with matching id
-
+        #?return storage
         return JSONResponse(
             message="ok",
             payload={
@@ -59,8 +57,22 @@ def get_storages():
             }
         ).to_json()
 
+    if request.method == 'PUT':
+        #?update storage information
+        body = request.get_json() #new info in request body
+        storage_name = body.get('storage_name', "")
+        validate_inputs({
+            "storage_name": validate_string(storage_name)
+        })
+
+        s.name = storage_name
+        s.description = body.get('description', '')
+        db.session.commit()
+
+        return JSONResponse(f"Storage-id-{storage_id} has been updated").to_json()
+
     if request.method == 'DELETE':
-        #delete storage with matching id
+        #?delete storage with matching id
         db.session.delete(s)
         db.session.commit()
         return JSONResponse(f"storage-id-{storage_id} has been deleted").to_json()
@@ -71,14 +83,10 @@ def get_storages():
 @user_required()
 def create_new_storage():
 
-    claims = get_jwt()
-    user = get_user_by_id(claims.get('user_id', None), company_required=True)
+    user = get_user_by_id(get_jwt().get('user_id', None), company_required=True)
     body = request.get_json(silent=True)
     storage_name = body['storage_name']
-
-    validate_inputs({
-        'storage_name': only_letters(storage_name, spaces=True)
-    })
+    
     try:
         new_storage = Storage(
             name = storage_name,
